feat(wip): untangle guests from R0 types
