+++ rust/common/src/lib.rs
@@ -1,5 +1,6 @@
 mod cache;
 mod cli;
+mod digest;
 mod guest;
 mod hashable;
 mod rpc;
@@ -8,6 +9,7 @@ pub mod verifier;
 
 pub use cache::InteriorMutabilityCache;
 pub use cli::{GlobalArgs, LogFormat};
+pub use digest::Digest;
 pub use guest::GuestElf;
 pub use hashable::Hashable;
 pub use rpc::Method;
