+++ rust/common/src/verifier.rs
@@ -1,3 +1,3 @@
 pub mod sealing;
-#[cfg(feature = "risc0")]
+
 pub mod zk_proof;
