+++ rust/common/src/verifier/zk_proof/mod.rs
@@ -1,0 +1,11 @@
+#[cfg(feature = "risc0")]
+mod risc0;
+
+#[cfg(feature = "risc0")]
+pub use risc0::*;
+
+#[cfg(feature = "sp1")]
+mod sp1;
+
+#[cfg(feature = "sp1")]
+pub use sp1::*;
