+++ rust/guest_wrapper/risc0_call_guest/Cargo.toml
@@ -7,7 +7,7 @@ edition = "2021"
 
 [dependencies]
 alloy-sol-types = { version = "=0.8.20", default-features = false }
-call_guest = { path = "../../services/call/guest" }
+call_guest = { path = "../../services/call/guest", features = ["risc0"] }
 risc0-zkvm = { version = "=2.0.0", default-features = false, features = ["std", "unstable"] }
 risc0-zkvm-platform = { version = "=2.0.1", features = [ "rust-runtime", "export-getrandom", "sys-getenv" ] }
 tokio = { version = "1.41.0", features = ["rt", "macros"] }
