+++ rust/services/call/guest/Cargo.toml
@@ -11,16 +11,21 @@ block_header = { workspace = true }
 call_common = { workspace = true }
 call_engine = { workspace = true }
 chain_client = { workspace = true }
-chain_common = { workspace = true }
-common = { workspace = true }
+chain_common = { workspace = true, default-features = false }
+common = { workspace = true, default-features = false  }
 mpt = { workspace = true }
 optimism = { workspace = true }
 revm = { workspace = true }
-risc0-zkvm = { workspace = true }
+risc0-zkvm = { workspace = true, optional = true }
 
 [dev-dependencies]
 call_engine = { workspace = true, features = ["testing"] }
 tokio = { workspace = true, features = ["rt", "macros"] }
+
+[features]
+default = ["risc0"]
+risc0 = ["dep:risc0-zkvm", "common/risc0", "chain_common/risc0"]
+sp1 = ["common/sp1", "chain_common/sp1"]
 
 [lints]
 workspace = true
