+++ rust/services/chain/common/Cargo.toml
@@ -10,18 +10,21 @@ block_header = { workspace = true, features = ["test-utils"] }
 block_trie = { workspace = true }
 bytemuck = { workspace = true }
 bytes = { workspace = true }
-common = { workspace = true }
+common = { workspace = true, default-features = false }
 derivative = { workspace = true }
 derive-new = { workspace = true }
 derive_more = { workspace = true }
 mpt = { workspace = true }
-risc0-zkp = { workspace = true }
-risc0-zkvm = { workspace = true }
+risc0-zkp = { workspace = true, optional = true }
+risc0-zkvm = { workspace = true, optional = true }
 serde = { workspace = true }
 serde_with = { workspace = true, features = ["macros", "hex"] }
 thiserror = { workspace = true }
 
 [features]
+default = ["risc0"]
+risc0 = ["dep:risc0-zkp", "dep:risc0-zkvm", "common/risc0"]
+sp1 = ["common/sp1"]
 testing = []
 
 [lints]
