+++ rust/services/chain/guest/Cargo.toml
@@ -7,11 +7,15 @@ edition = "2021"
 alloy-primitives = { workspace = true }
 block_header = { workspace = true }
 block_trie = { workspace = true }
-chain_common = { workspace = true }
+chain_common = { workspace = true, default-features = false  }
 common = { workspace = true }
-risc0-zkp = { workspace = true }
-risc0-zkvm = { workspace = true }
+risc0-zkvm = { workspace = true, optional = true }
 serde = { workspace = true }
+
+[features]
+default = ["risc0"]
+risc0 = ["dep:risc0-zkvm", "chain_common/risc0"]
+sp1 = ["chain_common/sp1"]
 
 [lints]
 workspace = true
