+++ rust/zkvm-benchmarks/benchmarks/src/benchmarks/accelerators.rs
@@ -6,8 +6,8 @@ mod rsa;
 
 pub fn benchmarks() -> Vec<Benchmark> {
     merge([
-        ("sha2", hash::sha2::benchmarks()),
-        ("keccak", hash::keccak::benchmarks()),
+        //("sha2", hash::sha2::benchmarks()),
+        //("keccak", hash::keccak::benchmarks()),
         ("rsa", rsa::benchmarks()),
     ])
 }
