+++ rust/zkvm-benchmarks/runner/sp1_guest/Cargo.toml
@@ -13,6 +13,6 @@ zkvm-benchmarks = { path = "../../benchmarks", features = ["sp1"] }
 sha2-v0-10-8 = { git = "https://github.com/sp1-patches/RustCrypto-hashes", package = "sha2", tag = "patch-sha2-0.10.8-sp1-4.0.0" }
 k256 = { git = "https://github.com/sp1-patches/elliptic-curves", tag = "patch-k256-13.4-sp1-4.1.0" }
 p256 = { git = "https://github.com/sp1-patches/elliptic-curves", tag = "patch-p256-13.2-sp1-4.1.0" }
-rsa = { git = "https://github.com/sp1-patches/RustCrypto-RSA/", rev = "481395d7ebb2f9910cd468e738ae92c7e40d405b" }
+rsa = { path = "../../../../../RustCrypto-RSA" }
 tiny-keccak = { git = "https://github.com/sp1-patches/tiny-keccak", tag = "patch-2.0.2-sp1-4.0.0" }
 ecdsa-core = { git = "https://github.com/sp1-patches/signatures", package = "ecdsa", tag = "patch-0.16.9-sp1-4.0.0" }
